Implemented BEM Template
Added cascading class structure to HTML document base, and added BEM-style class names to design.css and inquiries.css. Also applied Prettier from VSCode to the documents for improved readability. Now anyone can copy either of these html and css files, replace all of the instances of "design" or "inquiries" with the name of their page, and make changes that will apply strictly to the page they are working on.